Fixes to the construction of a GenericPolycone
This MR fixes the bug reported by Manuel Huber (Legend collaborator). The verified symptom is a segmentation fault during the construction of a GenericPolycone. Investigation showed that the problem was due to rounding in floating point comparisons used to decide whether new corners should be added to the reduced rz polygon (class ReducedPolycone).

This fix takes into account the vecgeom::kTolerance during those comparisons. A few functions are also improved, in terms of readability.

* Improved function ReducedPolycone::CalcPoIVectorFor2DPolygon()

* Remove spurious printout

* Improve checking variable name

* Improved CreateNewContour() used by GenericPolycone

* Improve readability for GetLineIntersection(line, zVal, poi)

* Account for tolerance in intersection validation

* Check y-coord as well during intersection validation

* Improve GetLineIntersection() functionality used in GenericPolycone

* GenericPolycone fix: take tolerance into account for floating point comparisons

* New simple GenericPolycone construction test

Based on a bug report by Manuel Huber (Legend collaboration)

* A few updates and fixes to v1.x series

This MR includes:
- Upgrade to built-in VecCore version, from 0.8.0 to 0.8.2 (latest tag)
- Fix the TestPolyhedra unit test
- Fix duplicate entries of libvecgeom.a while linking most tests (CUDA disabled,
  on OSX 14.6.1 with Clang 15.0.0) in client builds (e.g. Celeritas)
- Fix compilation error in LoggerTypes.cc (gcc 12.3.0) affecting Debug mode

* Fix polyhedra unit test

Modified Planes.h to include surface points into the Contains() condition,
and removed from TestPolyhedra unit test an unstable test:
  a Contains() test for a point in surface, which
  was failing in different ways for distinct platforms.

* Fix compilation error in LoggerTypes.cc (gcc 12.3.0)

* Attempt to fix duplicate entries of libvecgeom.a while linking most tests (CUDA disabled).

+++ source/LoggerTypes.cpp
@@ -20,7 +20,7 @@ char const *to_cstring(LogLevel lev)
   static const char *const data[] = {
       "debug", "diagnostic", "status", "info", "warning", "error", "critical",
   };
-  assert(size_t(lev) * sizeof(const char *) < sizeof(data));
+  assert((unsigned long)lev * sizeof(const char *) < sizeof(data));
   return data[static_cast<int>(lev)];
 }
